Adjusted numbers. Overall nerfs.
Particularly to the harvester, as further playtesting showed that it was able to penetrate plate armour with a steel harvester of good quality or higher, which is too powerful. The Vermillion Tide should be able to do that, however, so it's been rebalanced with this in mind.
Armour penetration values and crit has been adjusted for all weapons.
Crit was nerfed heavily as it was far too high to begin with. Pawns with equal melee stats would be almost guaranteed critting every single hit with a basic harvester.